Basic Prediction Functionality
Every certain amount of steps, the program will take the data of a step and the step following that one, then give the difference between the two. Over several simulations that data will be compiled and added to a large dataframe, which is used to train a Random Forest Regressor. The final simulation is currently used as a test simulation to see how the model fares against actual simulated data, but in this version the AI is given all of the data besides the value it's predicting (for example the X Position), which it then does for Y and Z positions. It will repeat this, predicting the positions and storing the values so they can be shown on the graph. The yellow line is the predicted trajectory of the object, while the red line is the actual simulated path. 

Next steps will be to remove the outliers for the time being after they escape the orbit (as I have a theory that very high positional values such as escape trajectories will result in heavily increased predicted positions due to weights on high positions being more influential to the model than low positions).