"Time chunk" AI prediction method (Unfinished)
I realized that I was displaying the wrong lines with the other prediction data, so I decided to start to make it more self-reliant. The AI will be fed 30 time steps of past positions in each direction, and it will then try to predict the most recent one. I'll use this predicted data as input by replacing the n-1 timestep with the predicted data, then dropping the n-n timestep (the first timestep in the dataframe in all positions). With the next update, I hope to loop this method to predict the orbital path. It's very likely it'll be inaccurate with the precision needed, but after I get it working I'll see how much I can tune it. 

This version's code is still quite messy, but it does have some preset simulation settings for ease of access. It will also display a prediction vs simulated data graph, and display the unsquared means squared error for each of the positional predictions. The MSE provided seems to be around 0.5% of the rough range of positions, but while that seems small I believe that the errors will compound over multiple predictions. If/when I get everything somewhat optimized, I'll likely label the code, clean up the unused portions, and then I'll see if I can create a release :)